tests/wpt/tests/webmessaging/with-ports/013.html

<!doctype html>
<title>loop in object in structured clone</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var x = {};
  x.foo = x;
  postMessage(x, '*', []);
  onmessage = this.step_func(function(e) {
    assert_equals(e.data, e.data.foo);
    this.done();
  });
});
</script>
